feat: serve OpenAPI HTML
