+++ packages/temba/create-api-for-quick-test.sh
@@ -32,8 +32,5 @@ echo "# API for quick testing
 This API is created on the fly to quickly test the Temba API against the local build in the \`dist\` folder.
 " > README.md
 
-# Copy the OpenAPI HTML page
-cp ../src/openapi/openapi.html ./openapi.html
-
 # Start the just created API
 npm run dev
