+++ packages/temba/src/index.ts
@@ -3,13 +3,12 @@ import { initConfig, type UserConfig } from './config'
 import type { IncomingMessage, ServerResponse } from 'http'
 import { createResourceHandler, handleNotFound, sendErrorResponse } from './resourceHandler'
 import { getHttpLogger, initLogger } from './log/logger'
-import { createOpenApiHandler } from './openapi/openapi'
+import { createOpenApiHandler, getOpenApiPaths } from './openapi'
 import { TembaError as TembaErrorInternal } from './requestInterceptor/TembaError'
 import { handleStaticFolder } from './staticFolder/staticFolder'
 import { getDefaultImplementations } from './implementations'
 import { setCorsHeaders } from './cors/cors'
 import { createRootUrlHandler } from './root/root'
-import { readFileSync } from 'node:fs'
 
 const removePendingAndTrailingSlashes = (url?: string) => (url ? url.replace(/^\/+|\/+$/g, '') : '')
 
@@ -23,14 +22,7 @@ const createServer = (userConfig?: UserConfig) => {
   const config = initConfig(userConfig)
 
   const rootPath = config.apiPrefix ? removePendingAndTrailingSlashes(config.apiPrefix) : ''
-  const openapiSpecPaths = [
-    `${rootPath ? `${rootPath}/` : ''}openapi.json`,
-    `${rootPath ? `${rootPath}/` : ''}openapi.yaml`,
-  ]
-  const openapiHtmlPaths = [
-    `${rootPath ? `${rootPath}/` : ''}openapi.html`,
-    `${rootPath ? `${rootPath}/` : ''}openapi`,
-  ]
+  const openapiPaths = getOpenApiPaths(rootPath)
   const { log, logLevel } = initLogger(process.env.LOG_LEVEL)
   const httpLogger = getHttpLogger(logLevel)
 
@@ -47,19 +39,6 @@ const createServer = (userConfig?: UserConfig) => {
           return handleOptionsRequest(req, res)
         }
 
-        if (openapiHtmlPaths.includes(requestUrl)) {
-          try {
-            const file = readFileSync('openapi.html')
-            setCorsHeaders(res)
-            res.writeHead(200, { 'Content-Type': 'text/html' })
-            res.end(file)
-          } catch {
-            res.writeHead(404)
-            res.end('Not found')
-          }
-          return
-        }
-
         if (config.staticFolder && !`${requestUrl}/`.startsWith(config.apiPrefix + '/')) {
           handleStaticFolder(req, res, () =>
             implementations.getStaticFileFromDisk(
@@ -68,12 +47,12 @@ const createServer = (userConfig?: UserConfig) => {
           )
         } else if (requestUrl === rootPath) {
           createRootUrlHandler(config)(req, res)
-        } else if (openapiSpecPaths.includes(requestUrl)) {
-          createOpenApiHandler(requestUrl.endsWith('.json') ? 'json' : 'yaml', config)(req, res)
+        } else if (openapiPaths.includes(requestUrl)) {
+          createOpenApiHandler(config, requestUrl, req.headers.host || '')(res)
         } else if (requestUrl.startsWith(rootPath)) {
           createResourceHandler(log, config)(req, res)
         } else {
-          handleNotFound(req, res)
+          handleNotFound(res)
         }
       }
