+++ packages/temba/src/openapi/html.ts
@@ -1,12 +1,13 @@
-<!doctype html>
-<!-- Important: must specify -->
+export const getOpenApiHtml = () => {
+  return `<!doctype html>
 <html>
   <head>
     <meta charset="utf-8" />
-    <!-- Important: rapi-doc uses utf8 characters -->
     <script type="module" src="https://unpkg.com/rapidoc/dist/rapidoc-min.js"></script>
   </head>
   <body>
     <rapi-doc spec-url="openapi.json"> </rapi-doc>
   </body>
 </html>
+  `
+}
